clean up initial implementation
