+++ derive_macros/src/derive/derive_unit_struct.rs
@@ -1,8 +1,8 @@
 use syn::{Attribute, Generics, Ident, Path};
 
 use crate::{
-    attr, bound,
-    derive::impl_tokens,
+    attr,
+    derive::{ImplStrategy, ImplValue},
     index::{Index, IndexAllocator, IndexValue},
     label::{label_or_ident, LabelValue},
     stream::stream_tag,
@@ -48,11 +48,6 @@ pub(crate) fn derive_unit_struct<'a>(
     generics: &Generics,
     attrs: &UnitStructAttrs,
 ) -> syn::Result<proc_macro2::TokenStream> {
-    let (impl_generics, ty_generics, _) = generics.split_for_impl();
-
-    let bound = parse_quote!(sval::Value);
-    let bounded_where_clause = bound::where_clause_with_bound(&generics, bound);
-
     let match_arm = stream_tag(
         quote!(_),
         attrs.tag(),
@@ -60,13 +55,9 @@ pub(crate) fn derive_unit_struct<'a>(
         attrs.index(),
     )?;
 
-    let tag = quote_optional_tag_owned(attrs.tag());
-
-    Ok(impl_tokens(
-        impl_generics,
+    ImplValue::new(Some(quote_optional_tag_owned(attrs.tag()))).quote_impl(
         ident,
-        ty_generics,
-        &bounded_where_clause,
+        generics,
         quote!({
             match self {
                 #match_arm
@@ -74,6 +65,5 @@ pub(crate) fn derive_unit_struct<'a>(
 
             sval::__private::result::Result::Ok(())
         }),
-        Some(tag),
-    ))
+    )
 }
