+++ derive_macros/src/lib.rs
@@ -7,10 +7,10 @@ extern crate core;
 
 mod attr;
 mod bound;
-mod codegen;
 mod derive;
 mod index;
 mod label;
+mod lifetime;
 mod stream;
 mod tag;
